// src/validation/models/ValidationType.ts
/*
 * Do not remove or alter the notices in this preamble.
 *
 * This software is owned by Worldline and may not be be altered, copied, reproduced, republished, uploaded, posted, transmitted or distributed in any way, without the prior written consent of Worldline.
 *
 * Copyright © 2024 Worldline and/or its affiliates.
 *
 * All rights reserved. License grant and user rights and obligations according to the applicable license agreement.
 *
 * Please contact Worldline for questions regarding license and user rights.
 */

import type { ValidationRule } from './ValidationRule'; // eslint-disable-line @typescript-eslint/no-unused-vars

/**
 * Enum contain all possible {@link ValidationRule} types.
 */
export enum ValidationType {
  ExpirationDate = 'EXPIRATIONDATE',
  EmailAddress = 'EMAILADDRESS',
  FixedList = 'FIXEDLIST',
  Iban = 'IBAN',
  Length = 'LENGTH',
  Luhn = 'LUHN',
  Range = 'RANGE',
  RegularExpression = 'REGULAREXPRESSION',
  Required = 'REQUIRED',
  Type = 'TYPE',
  TermsAndConditions = 'TERMSANDCONDITIONS',
}